Small bugfixes.

+++ PyPEEC/lib_mesher/png_mesher.py
@@ -25,7 +25,7 @@ def _get_idx_image(nx, ny, img, color):
         raise RunError("invalid image:  size is not compatible with the voxel structure")
 
     # get the color vector
-    color_tmp = np.array(color)
+    color_tmp = np.array(color, dtype=np.int64)
     color_tmp = np.expand_dims(color_tmp, axis=(0, 1))
     if not (color_tmp.shape == (1, 1, 4)):
         raise RunError("invalid color: colors should be a specified with for values")
